requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
Django==4.2.5
django-environ==0.11.2
django-ipware==5.0.0
django-redis==5.3.0
frozenlist==1.4.0
geographiclib==2.0
geoip2==4.7.0
geopy==2.4.0
hiredis==2.2.3
idna==3.4
maxminddb==2.4.0
more-itertools==9.1.0
multidict==6.0.4
openai==0.28.1
pycparser==2.21
pydantic==2.0.3
pydantic-core==2.3.0
python-ulid==1.1.0
redis==5.0.1
redis-om==0.2.1
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.1
types-pyOpenSSL==23.2.0.2
types-redis==4.6.0.7
typing-extensions==4.8.0
urllib3==2.0.5
yarl==1.9.2
